Ask the repository's models of the checks whose subject is the language
Three checks about the crossing from what a model states to where a run through
it is recorded were asked only of the corpus written against what the language
declares. That corpus reaches a construct about as often as it takes to declare
it, so what the crossing does where somebody wrote a model for its own reasons
was outside what they could see. It was asked of those models beside the
measurements instead, in a copy of the same questions carrying its own writing of
the projection they read through.

The three now take every model this repository carries. The models written for
shapes no corpus has stay beside each check: they are witnesses of a shape, not
substitutes for a population, and no corpus is asked to grow one.

The copy goes, and the question that was not a copy comes with it. Two of what it
asked are these checks' own. The third is about what the emitted tree has between
the copies both readings hold, which is what a reader wanting both readings of
one position has to recognise, and it belongs beside the check that reads the two
apart. Nothing here projects: where such a run begins, where it ends and what the
block it ends at belongs to are what the models are asked. Where a comparison of
the emitted tree was addressed, the projection the compiler already owns answers,
rather than a walk assembled again beside the question.

Reading a corpus as its manifest declares it takes a witness away, and what that
shows is that reachability over these models was never the contract it was being
read as. An arm about something a measure went without is reached only while a
model is short of that thing; a model made more adequate stops reaching it, and
holding the corpus still to keep the arm would be holding a model back from what
it is for. So the arms are on two lists now — the ones these models witness, and
the ones owed a witness of their own — and neither list is a note: an arm held to
be witnessed must be reached, and an arm recorded as owed must not be, so the day
a model reaches it the entry comes out. What such an arm is owed is a model
written to reach it through the analysis, which is a question about adequacy and
not about which models this repository carries.

The check said the difference between two sets and printed one of them, which is
what made the first reading of its failure wrong. It says the difference.
